refactor: 엔티티 디렉터리 이동
