+++ src/main/java/com/sscanner/team/points/entity/PaymentRecord.java
@@ -1,6 +1,8 @@
-package com.sscanner.team;
+package com.sscanner.team.points.entity;
 
+import com.sscanner.team.User;
 import com.sscanner.team.global.common.BaseEntity;
+import com.sscanner.team.products.entity.Product;
 import com.sscanner.team.util.UUIDConverter;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
