+++ src/main/java/com/sscanner/team/points/requestdto/PointUpdateRequestDto.java
@@ -1,7 +1,7 @@
 package com.sscanner.team.points.requestdto;
 
 import com.sscanner.team.User;
-import com.sscanner.team.UserPoint;
+import com.sscanner.team.points.entity.UserPoint;
 
 
 public record PointUpdateRequestDto(
