+++ src/main/java/com/sscanner/team/points/service/PaymentServiceImpl.java
@@ -1,9 +1,9 @@
 package com.sscanner.team.points.service;
 
-import com.sscanner.team.PaymentRecord;
-import com.sscanner.team.Product;
+import com.sscanner.team.points.entity.PaymentRecord;
+import com.sscanner.team.products.entity.Product;
 import com.sscanner.team.User;
-import com.sscanner.team.UserPoint;
+import com.sscanner.team.points.entity.UserPoint;
 import com.sscanner.team.global.exception.BadRequestException;
 import com.sscanner.team.global.exception.ExceptionCode;
 import com.sscanner.team.points.common.PointManager;
