+++ src/main/java/com/sscanner/team/products/entity/Product.java
@@ -1,14 +1,10 @@
-package com.sscanner.team;
+package com.sscanner.team.products.entity;
 
 import com.sscanner.team.global.common.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
-
-import java.time.Instant;
 
 @Getter
 @Entity
