+++ src/main/java/com/sscanner/team/products/entity/ProductQr.java
@@ -1,12 +1,8 @@
-package com.sscanner.team;
+package com.sscanner.team.products.entity;
 
 import com.sscanner.team.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
-
-import java.time.Instant;
 
 @Getter
 @Entity
