+++ src/main/java/com/sscanner/team/products/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.sscanner.team.products.repository;
 
-import com.sscanner.team.Product;
+import com.sscanner.team.products.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
